refactor(ai): seed wiring — immutable SeedScope off the general ship contract
Reworks how the decision seed travels (addresses PR review + a Codex design
pass). No behavior change; determinism preserved. Same green suite.

- SeedScope: an immutable, reproducible seed namespace. A per-agent root splits
  into independent child streams via Derive(streamId) — no RNG state, no runtime
  identity, so it replays across reconstructed episodes/processes. Centralizes all
  mixing (removes the hand-rolled Knuth spread in SolverBuffers and the ad-hoc
  per-consumer seed math).
- Off the general contract: DecisionSeed leaves IShipStatus (an AI-only concern
  was parked on the HUD/gunner-facing read interface) and rides the ShipControl
  commander-init bundle instead. Ship still carries the raw int (like teamNumber);
  the scope is materialized at BuildShipControl.
- Progressively narrowed, each owner names only its own streams: AICommander
  splits root into Nav/Strategy; Navigator derives MpcSampler; Brain derives
  UtilitySampler + Goal; GoalRunner materializes Patrol. Leaves take a plain
  int/uint (System.Random / Unity.Mathematics.Random) at the boundary, so the
  MPC solver stays decoupled from SeedScope.
- UnitService.Clear() now resets the agent index — Clear() is the episode-reset
  boundary, so an episode reset re-derives the same per-agent seeds without
  reconstructing the service (was a replay gap). New EpisodeReset test covers it.

Kept Navigator's explicit scope param (rejected the "read seed off IShipStatus"
shortcut — it conflicts with moving off IShipStatus and would hide an MPC
construction dependency in an unrelated status interface).

+++ src/Asteroids3D/Assets/Scripts/Editor/Tests/EditMode/MpcSolverTests.cs
@@ -53,7 +53,7 @@ namespace Tests.EditMode
         // so per-solve sampling noise cancels out.
         private State SolveTerminal(MpcInputs inputs, int warmup = 8, int average = 6)
         {
-            using var mpc = new Mpc(settings, dynamics, 0);
+            using var mpc = new Mpc(settings, dynamics, 0u);
             for (var i = 0; i < warmup; i++) mpc.Plan(in inputs);
 
             float2 posSum = default;
@@ -70,7 +70,7 @@ namespace Tests.EditMode
 
         private Control[] SolveSequence(int seed, MpcInputs inputs, int solves = 6)
         {
-            using var mpc = new Mpc(settings, dynamics, seed);
+            using var mpc = new Mpc(settings, dynamics, (uint)seed);
             for (var i = 0; i < solves; i++) mpc.Plan(in inputs);
             return (Control[])mpc.BestSequence.Clone();
         }
@@ -80,7 +80,7 @@ namespace Tests.EditMode
         // asserted here, at the raw candidate buffer, not on the planned output.
         private Control[] SolveCandidates(int seed, MpcInputs inputs)
         {
-            using var mpc = new Mpc(settings, dynamics, seed);
+            using var mpc = new Mpc(settings, dynamics, (uint)seed);
             mpc.Plan(in inputs);
             var solver = mpc.Solver;
             var count = solver.LastSampleCount * solver.LastHorizon;
